Successfull Bootloader Start
The current files should be succesfully be configured to be able to
compile code which can be loaded into disk and bochs will be able to
boot from it. This can be used as the starting pioint to start
implemenitng the bootloader. Simple instrucitons are in the README, more
work is needed.